Update README, refactor project structure
