+++ bot/private_chat.py
@@ -1,5 +1,5 @@
+from bot.chat_base import ChatBase
 from telegram import Message
-from chat_base import ChatBase
 
 class PrivateChat(ChatBase):
     def is_valid_message(self, message: Message) -> bool:
